backend/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
starlette>=0.27.0
python-multipart>=0.0.6
# Performance enhancements for uvicorn
uvloop>=0.17.0
httptools>=0.6.0

# API documentation
pydantic==1.10.22

# Database
sqlalchemy>=2.0.22
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.28.0

# Authentication & Security
cryptography>=41.0.4
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0
fastapi-limiter>=0.1.5
authlib>=1.2.0
itsdangerous>=2.1.2

# Email
emails>=0.6
jinja2>=3.1.2
aiosmtplib>=2.0.2

# QR code generation
qrcode>=7.4.2
Pillow>=10.0.1

# Redis for caching & Task Queue
redis>=5.0.1
hiredis>=2.2.3

# Background Task Queue
celery>=5.3.0
flower>=2.0.1

# Integration & Validation
jsonschema>=4.19.1
requests>=2.31.0
aiohttp>=3.12.0  # Added for SAP connector async HTTP requests
odoorpc>=0.9.0  # Added for Odoo integration
squareup>=21.0.0.20231030  # Square Python SDK for POS integration

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
# Removed pydantic-settings as it requires Pydantic V2
email-validator>=2.0.0
tenacity>=8.2.3
uuid>=1.30
